requirements.txt
Django==1.8.2
geolocation-python==0.2.0
requests==2.7.0
wheel==0.24.0
djangorestframework
